B--0--0.1--_indev--2025*2*5

+++ GithubAboutFile/DownloadFile.py
@@ -1,8 +1,8 @@
 from .FileExists import FileExists
-import ghapi.all as ghapi, base64
+import ghapi.all as ghapi, base64 # type: ignore
 
 
-def DownloadFile(access_token, owner, repo, file_path, local_path):
+def DownloadFile(access_token: str, owner: str, repo: str, file_path: str, local_path: str) -> None:
     """
     从 GitHub 仓库下载文件。
     
@@ -15,14 +15,14 @@ def DownloadFile(access_token, owner, repo, file_path, local_path):
     # 初始化 API 客户端
     api = ghapi.GhApi(owner=owner, repo=repo, token=access_token)
     
-    # if FileExists(access_token, owner, repo, file_path):
-    # 获取文件内容
-    response = api.repos.get_content(path=file_path) # type: ignore
-        
-    # 解码文件内容
-    decoded_content = base64.b64decode(response.content)
-        
-    # 将文件内容写入本地文件
-    with open(local_path, 'wb') as f:
-        f.write(decoded_content)
+    if FileExists(access_token, owner, repo, file_path):
+        # 获取文件内容
+        response = api.repos.get_content(path=file_path) # type: ignore
+            
+        # 解码文件内容
+        decoded_content = base64.b64decode(response.content) # type: ignore
+            
+        # 将文件内容写入本地文件
+        with open(local_path, 'wb') as f:
+            f.write(decoded_content)
